requirements.txt
Django>=4.0
djangorestframework
django-environ
psycopg2-binary
celery
django-redis
